Latte-Panda-Player/requirements.txt
python-mpv
paho-mqtt
netifaces
requests
